% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{build_network_graphs}
\alias{build_network_graphs}
\title{Build routable graphs for requested travel modes}
\usage{
build_network_graphs(osm_lines, modes = c("foot", "bicycle", "motorcar"))
}
\arguments{
\item{osm_lines}{An \code{sf} LINESTRING object with OpenStreetMap road attributes.}

\item{modes}{One or more of \code{foot}, \code{bicycle}, and \code{motorcar}.}
}
\value{
Named list of weighted \code{dodgr} graphs.
}
\description{
Build routable graphs for requested travel modes
}
