% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{create_app}
\alias{create_app}
\title{Construct the offline Shiny application}
\usage{
create_app(config = read_sampa_config(), equipment = NULL, graphs = NULL)
}
\arguments{
\item{config}{Project configuration.}

\item{equipment}{Optional equipment data.}

\item{graphs}{Optional network graphs.}
}
\value{
A shiny.appobj, usable with shiny::runApp or RStudio.
}
\description{
Construct the offline Shiny application
}
