% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_equipment}
\alias{validate_equipment}
\title{Validate normalized equipment records}
\usage{
validate_equipment(data, config = read_sampa_config())
}
\arguments{
\item{data}{Canonical equipment data.}

\item{config}{Project configuration.}
}
\value{
List containing annotated data, eligible records, quarantine, summary and issues.
}
\description{
Validate normalized equipment records
}
