% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{create_static_map}
\alias{create_static_map}
\title{Create a publication-oriented static proximity map}
\usage{
create_static_map(results, metric_id = NULL)
}
\arguments{
\item{results}{Long-form proximity results.}

\item{metric_id}{Metric to display.}
}
\value{
A ggplot object.
}
\description{
Create a publication-oriented static proximity map
}
